Fellner: Various fixes
1. There were various bugs when the lowest level model had primary
keys. Primary keys are unnecessary for the lowest level but should be
permitted for consistency.

2. For 3+ level models, the data was flattened for the nested
multilevel submodels. This was a waste of CPU time and, more
importantly, wrecked havoc with the flattening algorithm resulting in
observations placed in the wrong part of the covariance matrix.

3. The sign of the A matrix is correctly set based on whether the
shallow inverse is actually used or not instead of whether the shallow
inverse was potentially enabled (see signA).

4. For 3+ level models, the preorder tree traverse always aborted
prematurely.

5. We introduce a simpler scheme to cleanly reset state after testing
the depth of the shallow inverse.

Cleanly reset state after shallow inverse probe